create different component for recipe details
